Extract omicron-certificates from nexus-db-model
This will allow wicketd to share certificate validation for
user-provided certs during rack setup.
