+++ certificates/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "omicron-certificates"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[dependencies]
+openssl.workspace = true
+thiserror.workspace = true
+
+omicron-common.workspace = true
