+++ nexus/db-model/src/certificate.rs
@@ -8,90 +8,11 @@ use db_macros::Resource;
 use nexus_types::external_api::params;
 use nexus_types::external_api::views;
 use nexus_types::identity::Resource;
+use omicron_certificates::CertificateValidator;
 use omicron_common::api::external::Error;
-use openssl::asn1::Asn1Time;
-use openssl::pkey::PKey;
-use openssl::x509::X509;
 use uuid::Uuid;
 
-#[derive(Debug, thiserror::Error)]
-pub enum CertificateError {
-    #[error("Failed to parse certificate: {0}")]
-    BadCertificate(openssl::error::ErrorStack),
-
-    #[error("Certificate exists, but is empty")]
-    CertificateEmpty,
-
-    #[error("Certificate exists, but is expired")]
-    CertificateExpired,
-
-    #[error("Failed to parse private key")]
-    BadPrivateKey(openssl::error::ErrorStack),
-
-    #[error("Certificate and private key do not match")]
-    Mismatch,
-
-    #[error("Unexpected error: {0}")]
-    Unexpected(openssl::error::ErrorStack),
-}
-
-impl From<CertificateError> for Error {
-    fn from(error: CertificateError) -> Self {
-        use CertificateError::*;
-        match error {
-            BadCertificate(_) | CertificateEmpty | CertificateExpired
-            | Mismatch => Error::InvalidValue {
-                label: String::from("certificate"),
-                message: error.to_string(),
-            },
-            BadPrivateKey(_) => Error::InvalidValue {
-                label: String::from("private-key"),
-                message: error.to_string(),
-            },
-            Unexpected(_) => {
-                Error::InternalError { internal_message: error.to_string() }
-            }
-        }
-    }
-}
-
-fn validate_certs(input: &[u8]) -> Result<X509, CertificateError> {
-    let mut certs = X509::stack_from_pem(input)
-        .map_err(CertificateError::BadCertificate)?;
-    if certs.is_empty() {
-        return Err(CertificateError::CertificateEmpty);
-    }
-    let now =
-        Asn1Time::days_from_now(0).map_err(CertificateError::Unexpected)?;
-    for cert in &certs {
-        if cert.not_after() < now {
-            return Err(CertificateError::CertificateExpired);
-        }
-    }
-    // Return the first certificate in the chain (the leaf certificate)
-    // to use with verifying the private key.
-    Ok(certs.swap_remove(0))
-}
-
-fn validate_private_key(
-    cert: X509,
-    key: &[u8],
-) -> Result<(), CertificateError> {
-    let key = PKey::private_key_from_pem(key)
-        .map_err(CertificateError::BadPrivateKey)?;
-
-    // Verify the public key corresponding to this private key
-    // matches the public key in the certificate.
-    if !cert
-        .public_key()
-        .map_err(CertificateError::BadCertificate)?
-        .public_eq(&key)
-    {
-        return Err(CertificateError::Mismatch);
-    }
-
-    Ok(())
-}
+pub use omicron_certificates::CertificateError;
 
 /// Representation of x509 certificates used by services.
 #[derive(Queryable, Insertable, Clone, Selectable, Resource)]
@@ -125,8 +46,8 @@ impl Certificate {
         service: ServiceKind,
         params: params::CertificateCreate,
     ) -> Result<Self, CertificateError> {
-        let cert = validate_certs(&params.cert)?;
-        validate_private_key(cert, &params.key)?;
+        let validator = CertificateValidator::default();
+        validator.validate(&params.cert, &params.key)?;
 
         Ok(Self::new_unvalidated(silo_id, id, service, params))
     }
